Module Dirsp_proscript

Providing the signatures of cryptographic primitives needed by verified ProScript code.

There are other libraries, especially dirsp-proscript-mirage, that provide implementations for the cryptographic primitives.

exception Crypto_failure of string

exception Encoding_failure of string

module type PROSCRIPT_ENCODING = sig ... end

The ProScript.encoding interface for conversions back and forth between various encodings like hexadecimal.

module type PROSCRIPT_CRYPTO_ED25519 = sig ... end

The ProScript.crypto.ED25519 interface for the elliptic curve 25519 DSA.

module type PROSCRIPT_CRYPTO = sig ... end

The ProScript.crypto interface for core cryptographic primitives like SHA-256.

module type S = sig ... end

Signature of cryptographic primitives and low-level functions needed by verified ProScript code.